python-packages/green_light_ui/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "green_light_ui"
version = "0.0.1"
authors = [
  { name="Example Author", email="author@example.com" },
]
description = "The UI of the GreenLightServices (Road Accidents MLOps Project)"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies=[
"streamlit",
"pillow",
"fastapi",
"uvicorn",
"pandas",
"numpy",
"requests",
"pytest",
]
